requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.2
psycopg2-binary==2.9.7
redis==5.0.0
requests==2.31.0
python-dotenv==1.0.0
werkzeug==2.3.7
gunicorn==21.2.0
asyncio==3.4.3
aiohttp==3.8.5
psutil==5.9.5
jwt==1.3.1
PyJWT==2.8.0
cryptography==41.0.3
bcrypt==4.0.1
jsonschema==4.17.3
qrcode==7.4.2 